Put wrapping code into its own file
